# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "title": "Cargo transfer to ISS",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:8000",
    "basePath": "/",
    "paths": {
        "/api/cargo": {
            "get": {
                "description": "Возвращает все доступные грузы с опциональной фильтрацией по названию и диапазону цены",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Грузы"
                ],
                "summary": "Получить все грузы",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Название груза для фильтрации",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Нижний порог цены для фильтрации",
                        "name": "low_price",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Верхний порог цены для фильтрации",
                        "name": "high_price",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemes.GetAllCargosResponse"
                        }
                    }
                }
            }
        },
        "/api/cargo/": {
            "post": {
                "description": "Добавить новый груз",
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "Грузы"
                ],
                "summary": "Добавить груз",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Изображение груза",
                        "name": "image",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Название",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Английское название",
                        "name": "en_name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Категория",
                        "name": "category",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Цена",
                        "name": "price",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Вес",
                        "name": "weight",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Объем",
                        "name": "capacity",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Описание",
                        "name": "description",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/cargo/{cargo_id}": {
            "get": {
                "description": "Возвращает подробную информацию об одном грузе",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Грузы"
                ],
                "summary": "Получить один груз",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id груза",
                        "name": "cargo_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ds.Cargo"
                        }
                    }
                }
            },
            "put": {
                "description": "Изменить данные о грузе",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Грузы"
                ],
                "summary": "Изменить груз",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор груза",
                        "name": "cargo_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Изображение груза",
                        "name": "image",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Название",
                        "name": "name",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Английское название",
                        "name": "en_name",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Категория",
                        "name": "category",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "Цена",
                        "name": "price",
                        "in": "formData"
                    },
                    {
                        "type": "number",
                        "description": "Вес",
                        "name": "weight",
                        "in": "formData"
                    },
                    {
                        "type": "number",
                        "description": "Объем",
                        "name": "capacity",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Описание",
                        "name": "description",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "description": "Удаляет груз по id",
                "tags": [
                    "Грузы"
                ],
                "summary": "Удалить груз",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id груза",
                        "name": "cargo_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/cargo/{cargo_id}/add_to_flight": {
            "post": {
                "description": "Добавить выбранный груз в черновик полета",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Грузы"
                ],
                "summary": "Добавить груз в полет",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id груза",
                        "name": "cargo_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemes.AddToFlightResp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ds.Cargo": {
            "type": "object",
            "required": [
                "capacity",
                "category",
                "description",
                "en_name",
                "name",
                "price",
                "weight"
            ],
            "properties": {
                "capacity": {
                    "description": "m^3",
                    "type": "number"
                },
                "category": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "en_name": {
                    "type": "string"
                },
                "isDeleted": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "photo": {
                    "type": "string"
                },
                "price": {
                    "description": "Rubles",
                    "type": "integer"
                },
                "uuid": {
                    "type": "string"
                },
                "weight": {
                    "description": "kg",
                    "type": "number"
                }
            }
        },
        "schemes.AddToFlightResp": {
            "type": "object",
            "properties": {
                "cargo_count": {
                    "type": "integer"
                }
            }
        },
...
        "schemes.FlightShort": {
            "type": "object",
            "properties": {
                "cargo_count": {
                    "type": "integer"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "schemes.GetAllCargosResponse": {
            "type": "object",
            "properties": {
                "cargos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ds.Cargo"
                    }
                },
                "draft_flight": {
                    "$ref": "#/definitions/schemes.FlightShort"
                }
            }
        },
...
    }
}